make modulation v1
