+++ js/data.js
@@ -1,0 +1,58 @@
+import {createRandomIdFromRangeGenerator, getRandomPositiveInteger, getRandomArrayElement} from './util.js';
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+const DESCRIPTION = [
+  'Фото природы',
+  'Фото прибоя',
+  'Фото гор',
+  'Фото заката',
+  'Фото рассвета',
+  'Фото замка',
+  'Фото пустыни',
+  'Фото водопада',
+];
+
+const generatePhotoId = createRandomIdFromRangeGenerator(1, 25);
+const generatePhotoUrl = createRandomIdFromRangeGenerator(1, 25);
+const generateCommentId = createRandomIdFromRangeGenerator(1, 100);
+
+
+const createComents = () => ({
+  id: generateCommentId (),
+  avatar: `img/avatar-${getRandomPositiveInteger(1, 6)}.svg`,
+  message: getRandomArrayElement(COMMENTS),
+  name: getRandomArrayElement(NAMES),
+});
+
+
+const createFoto = () => ({
+  id: generatePhotoId(),
+  url: `photos/${generatePhotoUrl()}.jpg`,
+  description: getRandomArrayElement(DESCRIPTION),
+  likes: getRandomPositiveInteger(15, 200),
+  comments: Array.from({length: 2}, createComents),
+});
+
+
+const similarFotos = () => Array.from({length: 25}, createFoto);
+
+export {similarFotos};
